Address Gemini review: rename AdvancedSection → AudioSection, use Promise.allSettled
Rename AdvancedSection.svelte to AudioSection.svelte to match the tab
label and update the import in Settings.svelte. Switch GeneralSection's
loadSettings to Promise.allSettled so a hardware failure in get_microphones
(or any single invoke) doesn't block the rest of the section from loading.
